[Bronze V] Title: 이상한 기호, Time: 48 ms, Memory: 5492 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -6,27 +6,27 @@ class Program
     {
         string[] input = Console.ReadLine().Split();
         
-        int n1 = int.Parse(input[0]);
-        int n2 = int.Parse(input[1]);
+        long n1 = long.Parse(input[0]);
+        long n2 = long.Parse(input[1]);
         
-        int c1 = SumNum(n1,n2);
-        int c2 = MinusNum(n1,n2);
-        int c3 = ApplyNum(c1,c2);
+        long c1 = SumNum(n1,n2);
+        long c2 = MinusNum(n1,n2);
+        long c3 = ApplyNum(c1,c2);
         
         Console.WriteLine(c3.ToString());
     }
     
-    static int SumNum(int _n1, int _n2)
+    static long SumNum(long _n1, long _n2)
     {
         return _n1 + _n2; 
     }
     
-    static int MinusNum(int _n1, int _n2)
+    static long MinusNum(long _n1, long _n2)
     {
          return _n1 - _n2; 
     }
     
-    static int ApplyNum(int _n1, int _n2)
+    static long ApplyNum(long _n1, long _n2)
     {
         return _n1 * _n2;
     }
